requirements.txt
streamlit
google-generativeai
tavily-python
beautifulsoup4
pandas
openpyxl